[codegen] Checkpoint table parsing, again
I believe the parse functions are now working? with a few exceptions:

- we don't handle verification, at all
- we don't handle formats, at all
- we haven't added getters, yet

But! all in all I think this is okay, and I'm happy with our progress!

+++ font-tables/src/parse2.rs
@@ -171,8 +171,8 @@ impl<'a> Cursor<'a> {
     //}
 
     /// return the current position, or an error if we are out of bounds
-    pub(crate) fn position(&self) -> Result<u32, ReadError> {
-        self.data.check_in_bounds(self.pos).map(|_| self.pos as u32)
+    pub(crate) fn position(&self) -> Result<usize, ReadError> {
+        self.data.check_in_bounds(self.pos).map(|_| self.pos)
     }
 
     pub(crate) fn finish<T: TableInfo>(self, shape: T::Info) -> Result<TableRef<'a, T>, ReadError> {
